Return int8_t from system specific functions

+++ src/system/espidf/system.c
@@ -59,7 +59,7 @@ void z_task_wrapper(__z_task_arg *targ) {
 }
 
 /*------------------ Task ------------------*/
-int _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg) {
+int8_t _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg) {
     int ret = 0;
 
     __z_task_arg *z_arg = (__z_task_arg *)z_malloc(sizeof(__z_task_arg));
@@ -76,12 +76,12 @@ int _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), vo
     return ret;
 }
 
-int _z_task_join(_z_task_t *task) {
+int8_t _z_task_join(_z_task_t *task) {
     // Note: join not supported using FreeRTOS API
     return 0;
 }
 
-int _z_task_cancel(_z_task_t *task) {
+int8_t _z_task_cancel(_z_task_t *task) {
     vTaskDelete(task);
     return 0;
 }
@@ -93,24 +93,24 @@ void _z_task_free(_z_task_t **task) {
 }
 
 /*------------------ Mutex ------------------*/
-int _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, NULL); }
+int8_t _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, NULL); }
 
-int _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
+int8_t _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
 
-int _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
+int8_t _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
 
-int _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
+int8_t _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
 
-int _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
+int8_t _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
 
 /*------------------ Condvar ------------------*/
-int _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, NULL); }
+int8_t _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, NULL); }
 
-int _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
+int8_t _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
 
-int _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
+int8_t _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
 
-int _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
+int8_t _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
 #endif  // Z_MULTI_THREAD == 1
 
 /*------------------ Sleep ------------------*/
