+++ src/system/zephyr/system.c
@@ -56,7 +56,7 @@ K_THREAD_STACK_ARRAY_DEFINE(thread_stack_area, Z_THREADS_NUM, Z_PTHREAD_STACK_SI
 static int thread_index = 0;
 
 /*------------------ Task ------------------*/
-int _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg) {
+int8_t _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), void *arg) {
     _z_task_attr_t *lattr = NULL;
     _z_task_attr_t tmp;
     if (attr == NULL) {
@@ -68,9 +68,9 @@ int _z_task_init(_z_task_t *task, _z_task_attr_t *attr, void *(*fun)(void *), vo
     return pthread_create(task, lattr, fun, arg);
 }
 
-int _z_task_join(_z_task_t *task) { return pthread_join(*task, NULL); }
+int8_t _z_task_join(_z_task_t *task) { return pthread_join(*task, NULL); }
 
-int _z_task_cancel(_z_task_t *task) { return pthread_cancel(*task); }
+int8_t _z_task_cancel(_z_task_t *task) { return pthread_cancel(*task); }
 
 void _z_task_free(_z_task_t **task) {
     _z_task_t *ptr = *task;
@@ -79,24 +79,24 @@ void _z_task_free(_z_task_t **task) {
 }
 
 /*------------------ Mutex ------------------*/
-int _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, 0); }
+int8_t _z_mutex_init(_z_mutex_t *m) { return pthread_mutex_init(m, 0); }
 
-int _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
+int8_t _z_mutex_free(_z_mutex_t *m) { return pthread_mutex_destroy(m); }
 
-int _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
+int8_t _z_mutex_lock(_z_mutex_t *m) { return pthread_mutex_lock(m); }
 
-int _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
+int8_t _z_mutex_trylock(_z_mutex_t *m) { return pthread_mutex_trylock(m); }
 
-int _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
+int8_t _z_mutex_unlock(_z_mutex_t *m) { return pthread_mutex_unlock(m); }
 
 /*------------------ Condvar ------------------*/
-int _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, 0); }
+int8_t _z_condvar_init(_z_condvar_t *cv) { return pthread_cond_init(cv, 0); }
 
-int _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
+int8_t _z_condvar_free(_z_condvar_t *cv) { return pthread_cond_destroy(cv); }
 
-int _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
+int8_t _z_condvar_signal(_z_condvar_t *cv) { return pthread_cond_signal(cv); }
 
-int _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
+int8_t _z_condvar_wait(_z_condvar_t *cv, _z_mutex_t *m) { return pthread_cond_wait(cv, m); }
 #endif  // Z_MULTI_THREAD == 1
 
 /*------------------ Sleep ------------------*/
